Changed some math

+++ DACircularProgress/DACircularProgressView.h
@@ -13,6 +13,6 @@
 @property(nonatomic, strong) UIColor *trackTintColor;
 @property(nonatomic, strong) UIColor *progressTintColor;
 @property(nonatomic) BOOL roundedCorners;
-@property(nonatomic) float progress;
+@property(nonatomic) CGFloat progress;
 
 @end
